Fix possible cache construction race condition
The addition of a configurable cache (#76/#78) seems to have
introduced a possible race condition where multiple goroutines could
construct the caches. I discovered this while running tests for some
code which used `VerifyAccessToken` on multiple goroutines when using
the Go race detector.

My fix is to move the cache construction to the `New` functions;
however, this poses a problem as those functions do not include an
error return although the cache constructors could return an error. For
now, I have left the later constructor calls in order to punt returning
an error to a later call that returns an error. Alternatively, we could
either panic with an error (not ideal) or change the interface of the
`New` functions to include an error return.